Create blackjack.java
Java class blackjack that implements a function buildQortalAT to create an instance of a smart contract. The contract implements a blackjack game with the specified game duration (in minutes) and maximum number of players. The buildQortalAT function takes these two inputs, performs validity checks and returns the contract data as a byte array.